Stable implementation of Viewport ItemsControl
In this new implementation of AerialOBJ, we used IntemsControl instead of manually positioning canvas objects with boilerplate codes and such

+++ src/binstarjs03.AerialOBJ.WpfAppNew2/Components/ViewportObject.cs
@@ -4,6 +4,6 @@ namespace binstarjs03.AerialOBJ.WpfAppNew2.Components;
 public class ViewportObject
 {
     public int Size { get; set; }
-    public Point2<int> WorldPosition { get; set; }
+    public Point2<float> WorldPosition { get; set; }
     public Color Color { get; set; }
 }
